+++ examples/i2s-audio-out.rs
@@ -45,7 +45,6 @@ use cortex_m_rt::entry;
 use stm32_i2s_v12x::format::{Data16Frame16, FrameFormat};
 use stm32_i2s_v12x::{MasterClock, MasterConfig, Polarity};
 
-use stm32f4xx_hal::delay::Delay;
 use stm32f4xx_hal::i2c::I2c;
 use stm32f4xx_hal::i2s::I2s;
 use stm32f4xx_hal::nb::block;
@@ -100,7 +99,7 @@ fn main() -> ! {
     // The 86 MHz frequency can be divided to get a sample rate very close to 48 kHz.
     let clocks = rcc.cfgr.use_hse(8.MHz()).i2s_clk(86.MHz()).freeze();
 
-    let mut delay = Delay::new(cp.SYST, &clocks);
+    let mut delay = cp.SYST.delay(&clocks);
 
     let i2c = I2c::new(
         dp.I2C1,
