+++ examples/timer-syst.rs
@@ -15,7 +15,7 @@ use panic_halt as _;
 use cortex_m_rt::entry;
 use cortex_m_semihosting::hprintln;
 
-use hal::fugit::Error;
+use hal::timer::Error;
 use stm32f4xx_hal as hal;
 
 use crate::hal::{pac, prelude::*};
