+++ src/timer/hal_02.rs
@@ -9,12 +9,12 @@ use embedded_hal::{
     blocking::delay::{DelayMs, DelayUs},
     timer::{Cancel, CountDown, Periodic},
 };
-use fugit::{HertzU32 as Hertz, TimerDurationU32};
+use fugit::{ExtU32, HertzU32 as Hertz, TimerDurationU32};
 use void::Void;
 
 use super::{
-    delay::Wait, Channel, CounterHz, Delay, Error, Instance, Pins, Pwm, PwmChannel, SysCounterHz,
-    SysCounter, WithPwm,
+    delay::Wait, Channel, Counter, CounterHz, Delay, Error, FDelay, Instance, Pins, Pwm,
+    PwmChannel, PwmHz, SysCounter, SysCounterHz, WithPwm,
 };
 
 impl DelayUs<u32> for Delay<SYST> {
@@ -22,7 +22,7 @@ impl DelayUs<u32> for Delay<SYST> {
         // The SysTick Reload Value register supports values between 1 and 0x00FFFFFF.
         const MAX_RVR: u32 = 0x00FF_FFFF;
 
-        let mut total_rvr = us * (self.clk.raw() / 8_000_000);
+        let mut total_rvr = us * (self.clk.raw() / 1_000_000);
 
         while total_rvr != 0 {
             let current_rvr = if total_rvr <= MAX_RVR {
@@ -252,7 +252,7 @@ impl<TIM: Instance + WithPwm, const C: u8> embedded_hal::PwmPin for PwmChannel<T
     }
 }
 
-impl<TIM, P, PINS> embedded_hal::Pwm for Pwm<TIM, P, PINS>
+impl<TIM, P, PINS> embedded_hal::Pwm for PwmHz<TIM, P, PINS>
 where
     TIM: Instance + WithPwm,
     PINS: Pins<TIM, P>,
@@ -260,6 +260,116 @@ where
     type Channel = Channel;
     type Duty = u16;
     type Time = Hertz;
+
+    fn enable(&mut self, channel: Self::Channel) {
+        self.enable(channel)
+    }
+
+    fn disable(&mut self, channel: Self::Channel) {
+        self.disable(channel)
+    }
+
+    fn get_duty(&self, channel: Self::Channel) -> Self::Duty {
+        self.get_duty(channel)
+    }
+
+    fn set_duty(&mut self, channel: Self::Channel, duty: Self::Duty) {
+        self.set_duty(channel, duty)
+    }
+
+    /// If `0` returned means max_duty is 2^16
+    fn get_max_duty(&self) -> Self::Duty {
+        self.get_max_duty()
+    }
+
+    fn get_period(&self) -> Self::Time {
+        self.get_period()
+    }
+
+    fn set_period<T>(&mut self, period: T)
+    where
+        T: Into<Self::Time>,
+    {
+        self.set_period(period.into())
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> DelayUs<u32> for FDelay<TIM, FREQ> {
+    /// Sleep for `us` microseconds
+    fn delay_us(&mut self, us: u32) {
+        self.delay(us.micros()).unwrap()
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> DelayMs<u32> for FDelay<TIM, FREQ> {
+    /// Sleep for `ms` milliseconds
+    fn delay_ms(&mut self, ms: u32) {
+        self.delay(ms.millis()).unwrap()
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> DelayUs<u16> for FDelay<TIM, FREQ> {
+    /// Sleep for `us` microseconds
+    fn delay_us(&mut self, us: u16) {
+        self.delay((us as u32).micros()).unwrap()
+    }
+}
+impl<TIM: Instance, const FREQ: u32> DelayMs<u16> for FDelay<TIM, FREQ> {
+    /// Sleep for `ms` milliseconds
+    fn delay_ms(&mut self, ms: u16) {
+        self.delay((ms as u32).millis()).unwrap()
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> DelayUs<u8> for FDelay<TIM, FREQ> {
+    /// Sleep for `us` microseconds
+    fn delay_us(&mut self, us: u8) {
+        self.delay((us as u32).micros()).unwrap()
+    }
+}
+impl<TIM: Instance, const FREQ: u32> DelayMs<u8> for FDelay<TIM, FREQ> {
+    /// Sleep for `ms` milliseconds
+    fn delay_ms(&mut self, ms: u8) {
+        self.delay((ms as u32).millis()).unwrap()
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> Periodic for Counter<TIM, FREQ> {}
+
+impl<TIM: Instance, const FREQ: u32> CountDown for Counter<TIM, FREQ> {
+    type Time = TimerDurationU32<FREQ>;
+
+    fn start<T>(&mut self, timeout: T)
+    where
+        T: Into<Self::Time>,
+    {
+        self.start(timeout.into()).unwrap()
+    }
+
+    fn wait(&mut self) -> nb::Result<(), Void> {
+        match self.wait() {
+            Err(nb::Error::WouldBlock) => Err(nb::Error::WouldBlock),
+            _ => Ok(()),
+        }
+    }
+}
+
+impl<TIM: Instance, const FREQ: u32> Cancel for Counter<TIM, FREQ> {
+    type Error = Error;
+
+    fn cancel(&mut self) -> Result<(), Self::Error> {
+        self.cancel()
+    }
+}
+
+impl<TIM, P, PINS, const FREQ: u32> embedded_hal::Pwm for Pwm<TIM, P, PINS, FREQ>
+where
+    TIM: Instance + WithPwm,
+    PINS: Pins<TIM, P>,
+{
+    type Channel = Channel;
+    type Duty = u16;
+    type Time = TimerDurationU32<FREQ>;
 
     fn enable(&mut self, channel: Self::Channel) {
         self.enable(channel)
